Refac: Move query logic to custom repository
