+++ src/game-log/dto/game-record.dts.ts
@@ -1,0 +1,9 @@
+import { IsNumber } from 'class-validator';
+
+export class GameRecordDto {
+  @IsNumber()
+    total: number;
+
+  @IsNumber()
+    win: number;
+}
