+++ src/game-log/game-log.repository.ts
@@ -4,9 +4,37 @@ import GameLog from 'src/entities/game-log.entity';
 import GameOption from 'src/enums/mastercode/game-option.enum';
 import GameType from 'src/enums/mastercode/game-type.enum';
 import { EntityRepository, Repository } from 'typeorm';
+import { GameRecordDto } from './dto/game-record.dts';
 
 @EntityRepository(GameLog)
-export class GameLogRepository extends Repository<GameLog> {}
+export class GameLogRepository extends Repository<GameLog> {
+  private readonly logger = new Logger('GameLogRepository');
+
+  async findRecentGameLog(userSeq: number, limit: number): Promise<GameLog[] | GameLog> {
+    this.logger.debug('findRecentGameLog');
+    const ret = await this.find({
+      where: {
+        userSeq,
+        limit,
+      },
+    });
+    return ret;
+  }
+
+  async fundUserGameLog(userSeq: number): Promise<GameRecordDto> {
+    const [allGames, count] = await this.findAndCount({
+      where: [
+        { topUserSeq: userSeq },
+        { btmUserSeq: userSeq },
+      ],
+    });
+    const winGames = allGames.reduce((prev, wins) => {
+      if (wins.winnerSeq === userSeq) return prev + 1;
+      return prev;
+    }, 0);
+    return { total: count, win: winGames };
+  }
+}
 
 export class MockGameLogRepository {
   private readonly logger: Logger = new Logger('MockGameLogRepository');
@@ -76,6 +104,17 @@ export class MockGameLogRepository {
   async find(): Promise<GameLog[]> {
     this.logger.debug('find');
     return this.logs;
+  }
+
+  async findRecentGameLog(userSeq: number): Promise<GameLog[] | GameLog> {
+    return this.logs[0];
+  }
+
+  async findUserGameLog(userSeq:number): Promise<GameRecordDto> {
+    return {
+      total: 30,
+      win: 15,
+    };
   }
 
   /**
