+++ src/game-log/game-log.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
 import GameLog from 'src/entities/game-log.entity';
+import { GameRecordDto } from './dto/game-record.dts';
 import { GameLogRepository } from './game-log.repository';
 
 @Injectable()
@@ -16,26 +17,10 @@ export class GameLogService {
 
   async findRecentGameLog(userSeq: number, limit: number): Promise<GameLog[] | GameLog> {
     this.logger.debug('findRecentGameLog', userSeq, limit);
-    const ret = await this.gameLogRepository.find({
-      where: {
-        userSeq,
-        limit,
-      },
-    });
-    return ret;
+    return this.gameLogRepository.findRecentGameLog(userSeq, limit);
   }
 
-  async findUserGameLog(userSeq: number): Promise<{ total: number; win: number; }> {
-    const [allGames, count] = await this.gameLogRepository.findAndCount({
-      where: [
-        { topUserSeq: userSeq },
-        { btmUserSeq: userSeq },
-      ],
-    });
-    const winGames = allGames.reduce((prev, wins) => {
-      if (wins.winnerSeq === userSeq) return prev + 1;
-      return prev;
-    }, 0);
-    return { total: count, win: winGames };
+  async findUserGameLog(userSeq: number): Promise<GameRecordDto> {
+    return this.gameLogRepository.fundUserGameLog(userSeq);
   }
 }
